no message
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@98 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ applications/mp4client/main.c
@@ -197,7 +197,7 @@ static void UpdateRTInfo()
 	if (!rti.process_memory) rti.process_memory = (u32) (memory_at_gpac_startup-rti.physical_memory_avail);
 
 	sprintf(szMsg, "FPS %02.2f - CPU %02d (%02d) - Mem %d kB", 
-		gf_term_get_framerate(term, 0), rti.total_cpu_usage, rti.process_cpu_usage, rti.process_memory / 1024);
+		gf_term_get_framerate(term, 0), rti.total_cpu_usage, rti.process_cpu_usage, (u32) (rti.process_memory / 1024) );
 
 	evt.type = GF_EVT_SET_CAPTION;
 	evt.caption.caption = szMsg;
